Projects/3. RAG Q&A Conversation/requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
langchain_core
langchain-openai
streamlit
langchain_groq
langchain-text-splitters
sentence_transformers
langchain_huggingface
openai
faiss-cpu
chromadb
duckdb
pypdf
pymupdf
tf-keras
langchain_chroma